Create README.md
Bitcoin is Where All Money will Flow in the NEXT 50 Years. All of it! With projects like Lightning and SegWit 2x in motion, it is unlikely that any other 'Crypto' or Fiat Currency will beat Bitcoin in terms of the 'Network Effects' of it being a 1) Superior store of Value, 2) Unit of Account, and 3) Medium of Exchange. https://muchiniwam.com/bitcoin